refactor: request dto 생성 및 패키지 이동
