+++ app/src/main/java/woowacourse/shopping/data/dto/request/AddCartRequestDto.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.data.dto.request
+
+import com.google.gson.annotations.SerializedName
+
+data class AddCartRequestDto(
+    @SerializedName("productId")
+    val productId: Long,
+)
